# openapi.yaml
openapi: 3.0.3
info:
  title: API da Oficina Mecanica
  description: |
    Back-end monolitico do MVP da oficina. Status da OS - Recebida, Em diagnostico, Aguardando aprovacao, Em execucao, Finalizada, Entregue; Cancelada e 7o status decidido no ADR-008.

    Autorizacao: todo endpoint fora da tag Acompanhamento e do login exige JWT e responde 401 sem token valido. Para nao repetir a mesma resposta em 36 operacoes, o 401 e declarado nas operacoes de referencia (os quatro verbos de Clientes, as listagens dos quatro cadastros, a listagem de Ordens de Servico, a de pendencias de pecas e a metrica de tempo medio); a ausencia da declaracao nas demais nao significa acesso liberado. O 401 do login e outro: ele e publico, nao envolve JWT nenhum, e responde credencial invalida.

...
  version: 1.0.0
servers:
  - {url: 'http://localhost:8080/api/v1', description: ambiente local via docker-compose}
...
security:
  - jwtAdministrativo: []
paths:
...
  /ordens-servico/{id}:
    get:
      tags: [OrdensServico]
      summary: Detalhar Ordem de Servico
      description: Detalhe da OS com itens, versoes de Orcamento e Transicoes de status.
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      responses:
        '200': {description: Detalhe da OS, content: {application/json: {schema: {$ref: '#/components/schemas/OrdemServico'}}}}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
...
  /pecas:
    post:
      tags: [Pecas]
      summary: Cadastrar peca
      description: 'Peca ou Insumo (unidade de medida), com preco e Estoque minimo. Nasce com Saldo em estoque zero: o saldo entra pelo caminho modelado, que e a Entrada de estoque.'
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PecaRequest'}}}}
      responses:
        '201': {description: Peca cadastrada, content: {application/json: {schema: {$ref: '#/components/schemas/Peca'}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
    get:
      tags: [Pecas]
      summary: Listar pecas
      description: 'O filtro abaixoDoMinimo entrega a visao Pecas abaixo do Estoque minimo, que o Estoquista consulta. Nao e o Alerta de reposicao, que e aviso automatico e ficou fora do MVP (ADR-013). Lista apenas pecas ativas: as removidas logicamente nao aparecem (ADR-014).'
      parameters:
        - {name: abaixoDoMinimo, in: query, required: false, schema: {type: boolean}}
      responses:
        '200': {description: Lista de pecas com saldo, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Peca'}}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
        '401': {$ref: '#/components/responses/NaoAutorizado'}
  /pecas/{id}:
    get:
      tags: [Pecas]
      summary: Detalhar peca
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      responses:
        '200': {description: Peca, content: {application/json: {schema: {$ref: '#/components/schemas/Peca'}}}}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
    put:
      tags: [Pecas]
      summary: Alterar peca
      description: 'Altera os dados de cadastro da Peca: nome, unidade de medida, preco e Estoque minimo. O saldo NAO muda aqui: ele so se move por Entrada de estoque, Baixa de estoque e Devolucao; a Reserva de pecas move a quantidade reservada, nao o saldo.'
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PecaRequest'}}}}
      responses:
        '200': {description: Peca alterada, content: {application/json: {schema: {$ref: '#/components/schemas/Peca'}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
    delete:
      tags: [Pecas]
      summary: Remover peca
      description: 'Recusado (409) se houver OS em RECEBIDA, EM_DIAGNOSTICO, AGUARDANDO_APROVACAO ou EM_EXECUCAO ou Reserva de peca ATIVA referenciando a peca. A remocao e LOGICA: o registro passa a inativo, some das listagens e do detalhe, e permanece no banco porque o historico de Ordens de Servico o referencia por identidade (ADR-014).'
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      responses:
        '204': {description: Peca removida}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
        '409': {$ref: '#/components/responses/ConflitoDeEstado'}
  /pecas/{id}/entradas-estoque:
    post:
      tags: [Pecas]
      summary: Registrar entrada de estoque
      description: Executado pelo Estoquista. Recebimento de compra; aumenta o Saldo em estoque.
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EntradaEstoque'}}}}
      responses:
        '200': {description: 'entrada registrada; devolve a Peca com o saldo atualizado. Responde 200 e nao 201 porque nenhum recurso novo nasce: o ADR-013 recusou a tabela de movimentacao, e a Entrada e registrada pelo seu efeito no saldo', content: {application/json: {schema: {$ref: '#/components/schemas/Peca'}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
  /ordens-servico/{id}/pecas-reservadas:
    get:
      tags: [Pecas]
      summary: Consultar pecas reservadas da OS
      description: 'Modelo de leitura Pecas reservadas por OS, que o Mecanico consulta antes de retirar. E por aqui que ele obtem o id de cada Reserva de pecas, que a retirada e a devolucao exigem na requisicao (ADR-013).'
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      responses:
        '200': {description: 'reservas da OS', content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ReservaPeca'}}}}}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
  /ordens-servico/{id}/pecas/retirada:
    post:
      tags: [Pecas]
      summary: Retirar pecas reservadas
      description: 'Executado pelo Mecanico, com a OS em EM_EXECUCAO. Consome a reserva INTEIRA e registra a Baixa de estoque, que acontece AQUI e nao na conclusao da execucao (ADR-013), para o Saldo em estoque nao mentir durante a execucao. Retirada parcial nao existe: a reserva alvo e identificada pelo seu id e passa a CONSUMIDA por completo (ADR-013). Reserva que nao esteja ATIVA e recusada com 409.'
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReservasRequest'}}}}
      responses:
        '200': {description: 'baixa registrada; reservas passam a CONSUMIDA', content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ReservaPeca'}}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
        '409': {$ref: '#/components/responses/ConflitoDeEstado'}
  /ordens-servico/{id}/pecas/devolucao:
    post:
      tags: [Pecas]
      summary: Devolver peca nao utilizada
      description: 'Executado pelo Mecanico ou Estoquista, com a OS em EM_EXECUCAO, FINALIZADA ou ENTREGUE. Os tres estados entram porque a reserva precisa ser alcancavel ate o fim da vida da OS: em qualquer status que nao aceite a devolucao, uma reserva ATIVA fica orfa e trava o saldo para sempre (ADR-013). Unico estorno do MVP, manual e por decisao registrada (ADR-013), espelhando a conferencia fisica antes de devolver ao saldo. **Reserva que ja esteja DEVOLVIDA e recusada com 409**, senao devolver duas vezes a mesma reserva somaria saldo duas vezes e criaria peca do nada. O efeito depende da situacao da reserva IDENTIFICADA na requisicao: reserva CONSUMIDA ja tirou a peca do saldo na Baixa, entao a devolucao REPOE o Saldo em estoque; reserva ATIVA nunca saiu do saldo, so da quantidade reservada, entao a devolucao LIBERA a reserva sem mexer no saldo. Aumentar o saldo nos dois casos criaria peca do nada e quebraria a invariante de que o saldo precisa ser verdadeiro. E por depender da situacao que o alvo e a reserva e nao a peca: a mesma Peca pode ter duas reservas na mesma OS em situacoes diferentes. Devolucao parcial nao existe: a reserva passa a DEVOLVIDA por completo (ADR-013).'
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReservasRequest'}}}}
      responses:
        '200': {description: 'devolucao registrada; reservas passam a DEVOLVIDA', content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/ReservaPeca'}}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
        '409': {$ref: '#/components/responses/ConflitoDeEstado'}
  /ordens-servico/{id}/pecas/faltas:
    post:
      tags: [Pecas]
      summary: Registrar falta de peca
      description: 'Executado pelo Mecanico, com a OS em EM_EXECUCAO, quando identifica na execucao que uma peca do escopo aprovado nao esta disponivel. A OS SEGUE em EM_EXECUCAO e a falta vira pendencia: nao existe status Aguardando pecas (ADR-013). E o segundo caminho de entrada do modelo de leitura Pendencias de pecas por OS; o primeiro e automatico, na Reserva de pecas com estoque insuficiente.'
      parameters: [{$ref: '#/components/parameters/IdRecurso'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FaltaPeca'}}}}
      responses:
        '201': {description: 'pendencia registrada', content: {application/json: {schema: {$ref: '#/components/schemas/PendenciaPeca'}}}}
        '400': {$ref: '#/components/responses/RequisicaoInvalida'}
        '404': {$ref: '#/components/responses/NaoEncontrado'}
        '409': {$ref: '#/components/responses/ConflitoDeEstado'}
...
components:
  securitySchemes:
    jwtAdministrativo:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT das APIs administrativas; obtido em /auth/login. Toda rota administrativa responde 401 sem token valido.
  parameters:
    IdRecurso:
      {name: id, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    RequisicaoInvalida:
      description: Dados invalidos (inclui Documento ou Placa reprovados na validacao de dominio)
      content: {application/json: {schema: {$ref: '#/components/schemas/Erro'}}}
    NaoAutorizado:
      description: JWT ausente, invalido ou expirado
      content: {application/json: {schema: {$ref: '#/components/schemas/Erro'}}}
...
    NaoEncontrado:
      description: 'Recurso inexistente. Nas rotas de acompanhamento o 404 e UNIFORME (ADR-007): codigo inexistente e codigo de outra OS respondem igual, nunca distinguindo "nao existe" de "nao e seu"'
      content: {application/json: {schema: {$ref: '#/components/schemas/Erro'}}}
    ConflitoDeEstado:
      description: 'Operacao viola a maquina de estados da OS ou uma invariante de dominio (ex. Documento ou Placa ja cadastrados, cadastro com OS em andamento, reserva fora da situacao que a operacao exige)'
      content: {application/json: {schema: {$ref: '#/components/schemas/Erro'}}}
  schemas:
    Erro:
      type: object
      required: [codigo, mensagem]
      properties:
        codigo: {type: string, example: TRANSICAO_INVALIDA}
        mensagem: {type: string}
...
    StatusOrdemServico:
      type: string
      description: os 6 status do fluxo de atendimento; CANCELADA e 7o status decidido no ADR-008
      enum: [RECEBIDA, EM_DIAGNOSTICO, AGUARDANDO_APROVACAO, EM_EXECUCAO, FINALIZADA, ENTREGUE, CANCELADA]
    Dinheiro:
      type: object
      required: [valor, moeda]
      properties:
        valor: {type: number, minimum: 0, maximum: 9999999999.99, description: 'duas casas decimais; o teto espelha a coluna NUMERIC(12,2) que persiste o valor. A validacao das casas mora no dominio, e nao em multipleOf, porque multipleOf sobre numero de ponto flutuante rejeita valores legitimos de duas casas'}
        moeda: {type: string, enum: [BRL]}
...
    PecaRequest:
      type: object
      required: [nome, preco, estoqueMinimo]
      properties:
        nome: {type: string, minLength: 1, maxLength: 120, example: Filtro de oleo}
        unidadeMedida: {type: string, minLength: 1, maxLength: 20, default: unidade, description: 'para Insumo (ex. litro); pecas seriadas usam unidade. Omitido, o SERVIDOR aplica o default antes de persistir, porque a coluna e NOT NULL'}
        preco: {$ref: '#/components/schemas/Dinheiro'}
        estoqueMinimo: {type: integer, minimum: 0, maximum: 1000000}
    Peca:
      allOf:
        - {$ref: '#/components/schemas/PecaRequest'}
        - type: object
          required: [id, saldoEmEstoque, quantidadeReservada]
          properties:
            id: {type: string, format: uuid}
            saldoEmEstoque: {type: integer, minimum: 0, description: 'invariante de dominio, nunca negativo'}
            quantidadeReservada: {type: integer, minimum: 0}
    EntradaEstoque:
      type: object
      required: [quantidade]
      properties:
        quantidade: {type: integer, minimum: 1, maximum: 1000000, description: 'teto operacional, menor que o da coluna de proposito, porque esta quantidade se SOMA ao Saldo em estoque existente'}
...
    ReservasRequest:
      type: object
      required: [reservas]
      description: 'Identifica as Reservas de pecas alvo da operacao. A reserva e o alvo, e nao o par peca mais quantidade, porque ela tem ciclo de vida proprio (ativa, consumida, devolvida) e a mesma Peca pode ter mais de uma reserva na mesma OS, em situacoes diferentes, quando um Reparo adicional a inclui de novo. A operacao move a reserva INTEIRA de situacao: nao existe retirada nem devolucao parcial (ADR-013)'
      properties:
        reservas:
          type: array
          minItems: 1
          items:
            type: object
            required: [reservaId]
            properties:
              reservaId: {type: string, format: uuid, description: 'identificador da Reserva de pecas, obtido na consulta de pecas reservadas da OS'}
    ReservaPeca:
      type: object
      required: [id, pecaId, quantidade, situacao]
      properties:
        id: {type: string, format: uuid}
        pecaId: {type: string, format: uuid}
        nomePeca: {type: string}
        quantidade: {type: integer, minimum: 1}
        situacao: {type: string, enum: [ATIVA, CONSUMIDA, DEVOLVIDA]}
        criadaEm: {type: string, format: date-time}
    ItemOrcamento:
      type: object
      required: [tipo, nome, valor]
      properties:
        tipo: {type: string, enum: [SERVICO, PECA], description: 'o CDC art. 40 exige o orcamento discriminando mao de obra e materiais; e essa a separacao'}
        nome: {type: string, maxLength: 120, description: 'rotulo ATUAL do Servico ou da Peca no catalogo; a imutabilidade da versao cobre valores e composicao, nao o rotulo de exibicao'}
        quantidade: {type: integer, minimum: 1, description: 'presente nos itens de peca'}
        valor: {allOf: [{$ref: '#/components/schemas/Dinheiro'}], description: 'valor unitario copiado como snapshot na inclusao'}
    FaltaPeca:
      type: object
      required: [pecaId, quantidadeFaltante]
      properties:
        pecaId: {type: string, format: uuid}
        quantidadeFaltante: {type: integer, minimum: 1, maximum: 1000000}
    PendenciaPeca:
      type: object
      required: [id, ordemServicoId, pecaId, quantidadeFaltante, detectadaEm]
      properties:
        id: {type: string, format: uuid}
        ordemServicoId: {type: string, format: uuid}
        pecaId: {type: string, format: uuid}
        nomePeca: {type: string}
        quantidadeFaltante: {type: integer, minimum: 1}
        detectadaEm: {type: string, format: date-time}
        resolvidaEm: {type: string, format: date-time, nullable: true, description: 'reservada para quando a resolucao da pendencia for implementada; fora do MVP (ADR-013), entao hoje toda pendencia carregada esta aberta'}
    ItemServico:
      type: object
      required: [servicoId]
      properties:
        servicoId: {type: string, format: uuid}
        valorMaoDeObraSnapshot: {allOf: [{$ref: '#/components/schemas/Dinheiro'}], readOnly: true, description: 'copiado do catalogo pelo servidor na inclusao; ignorado se enviado'}
        versaoOrcamento: {type: integer, minimum: 1, readOnly: true, description: 'versao do Orcamento que introduziu este item (ADR-006)'}
    ItemPeca:
      type: object
      required: [pecaId, quantidade]
      properties:
        pecaId: {type: string, format: uuid}
        quantidade: {type: integer, minimum: 1, maximum: 1000000}
        precoSnapshot: {allOf: [{$ref: '#/components/schemas/Dinheiro'}], readOnly: true, description: 'copiado do catalogo pelo servidor na inclusao; ignorado se enviado'}
        versaoOrcamento: {type: integer, minimum: 1, readOnly: true, description: 'versao do Orcamento que introduziu este item (ADR-006)'}
...
    Orcamento:
      type: object
      required: [versao, situacao, total, validadeDias]
      properties:
        versao: {type: integer, minimum: 1}
        validadeDias: {type: integer, minimum: 1, default: 10, description: 'validade padrao de 10 dias (CDC art. 40)'}
        situacao: {type: string, enum: [PENDENTE, APROVADO, REPROVADO]}
        total: {$ref: '#/components/schemas/Dinheiro'}
        itens: {type: array, items: {$ref: '#/components/schemas/ItemOrcamento'}, description: 'valor discriminado do escopo coberto por esta versao; e o que o Cliente le antes de aprovar. Item introduzido por versao ANTERIOR ja reprovada nao entra, porque o reparo adicional recusado nao volta ao escopo (ADR-008); a propria versao nunca se autoexclui, entao uma versao reprovada exibe os itens que ela introduziu'}
        dataEnvio: {type: string, format: date-time}
        dataResposta: {type: string, format: date-time}
        descricao: {type: string, maxLength: 500, description: preenchida quando a versao nasce de um Reparo adicional; e o que o Cliente le antes de autorizar}
    TransicaoStatus:
      type: object
      required: [paraStatus, dataHora]
      properties:
        deStatus: {$ref: '#/components/schemas/StatusOrdemServico'}
        paraStatus: {$ref: '#/components/schemas/StatusOrdemServico'}
        dataHora: {type: string, format: date-time, description: timestamp que viabiliza o Tempo medio de execucao}
    OrdemServicoSummary:
      type: object
      required: [id, status, clienteId, veiculoId]
      properties:
        id: {type: string, format: uuid}
        status: {$ref: '#/components/schemas/StatusOrdemServico'}
        clienteId: {type: string, format: uuid}
        veiculoId: {type: string, format: uuid}
        criadaEm: {type: string, format: date-time}
    OrdemServico:
      allOf:
        - {$ref: '#/components/schemas/OrdemServicoSummary'}
        - type: object
          properties:
            codigoAcompanhamento: {type: string}
            relatoDoProblema: {type: string}
            itensServico: {type: array, items: {$ref: '#/components/schemas/ItemServico'}}
            itensPeca: {type: array, items: {$ref: '#/components/schemas/ItemPeca'}}
            orcamentos: {type: array, description: versoes imutaveis apos envio (ADR-006), items: {$ref: '#/components/schemas/Orcamento'}}
            transicoesStatus: {type: array, items: {$ref: '#/components/schemas/TransicaoStatus'}}
...